+++ pkg/cmd/instancetype_test.go
@@ -1,0 +1,36 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+package cmd
+
+import (
+	"testing"
+
+	"github.com/nirvana-labs/nirvana-cli/internal/mocktest"
+)
+
+func TestInstanceTypesList(t *testing.T) {
+	t.Skip("Mock server tests are disabled")
+	t.Run("regular flags", func(t *testing.T) {
+		mocktest.TestRunMockTestWithFlags(
+			t,
+			"--api-key", "string",
+			"instance-types", "list",
+			"--max-items", "10",
+			"--cursor", "cursor",
+			"--limit", "10",
+		)
+	})
+}
+
+func TestInstanceTypesGet(t *testing.T) {
+	t.Skip("Mock server tests are disabled")
+	t.Run("regular flags", func(t *testing.T) {
+		mocktest.TestRunMockTestWithFlags(
+			t,
+			"--api-key", "string",
+			"instance-types", "get",
+			"--region", "us-sva-2",
+			"--name", "n1-standard-8",
+		)
+	})
+}