+++ pkg/cmd/membership.go
@@ -1,0 +1,163 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+package cmd
+
+import (
+	"context"
+	"fmt"
+	"os"
+
+	"github.com/nirvana-labs/nirvana-cli/internal/apiquery"
+	"github.com/nirvana-labs/nirvana-cli/internal/requestflag"
+	"github.com/nirvana-labs/nirvana-go"
+	"github.com/nirvana-labs/nirvana-go/option"
+	"github.com/nirvana-labs/nirvana-go/organizations"
+	"github.com/tidwall/gjson"
+	"github.com/urfave/cli/v3"
+)
+
+var organizationsMembershipsList = cli.Command{
+	Name:    "list",
+	Usage:   "List all memberships for an organization",
+	Suggest: true,
+	Flags: []cli.Flag{
+		&requestflag.Flag[string]{
+			Name:     "organization-id",
+			Required: true,
+		},
+		&requestflag.Flag[string]{
+			Name:      "cursor",
+			Usage:     "Pagination cursor returned by a previous request",
+			QueryPath: "cursor",
+		},
+		&requestflag.Flag[int64]{
+			Name:      "limit",
+			Usage:     "Maximum number of items to return",
+			Default:   10,
+			QueryPath: "limit",
+		},
+		&requestflag.Flag[int64]{
+			Name:  "max-items",
+			Usage: "The maximum number of items to return (use -1 for unlimited).",
+		},
+	},
+	Action:          handleOrganizationsMembershipsList,
+	HideHelpCommand: true,
+}
+
+var organizationsMembershipsGet = cli.Command{
+	Name:    "get",
+	Usage:   "Get details about an organization membership",
+	Suggest: true,
+	Flags: []cli.Flag{
+		&requestflag.Flag[string]{
+			Name:     "organization-id",
+			Required: true,
+		},
+		&requestflag.Flag[string]{
+			Name:     "membership-id",
+			Required: true,
+		},
+	},
+	Action:          handleOrganizationsMembershipsGet,
+	HideHelpCommand: true,
+}
+
+func handleOrganizationsMembershipsList(ctx context.Context, cmd *cli.Command) error {
+	client := nirvana.NewClient(getDefaultRequestOptions(cmd)...)
+	unusedArgs := cmd.Args().Slice()
+	if !cmd.IsSet("organization-id") && len(unusedArgs) > 0 {
+		cmd.Set("organization-id", unusedArgs[0])
+		unusedArgs = unusedArgs[1:]
+	}
+	if len(unusedArgs) > 0 {
+		return fmt.Errorf("Unexpected extra arguments: %v", unusedArgs)
+	}
+
+	params := organizations.MembershipListParams{}
+
+	options, err := flagOptions(
+		cmd,
+		apiquery.NestedQueryFormatBrackets,
+		apiquery.ArrayQueryFormatComma,
+		EmptyBody,
+		false,
+	)
+	if err != nil {
+		return err
+	}
+
+	format := cmd.Root().String("format")
+	transform := cmd.Root().String("transform")
+	if format == "raw" {
+		var res []byte
+		options = append(options, option.WithResponseBodyInto(&res))
+		_, err = client.Organizations.Memberships.List(
+			ctx,
+			cmd.Value("organization-id").(string),
+			params,
+			options...,
+		)
+		if err != nil {
+			return err
+		}
+		obj := gjson.ParseBytes(res)
+		return ShowJSON(os.Stdout, "organizations:memberships list", obj, format, transform)
+	} else {
+		iter := client.Organizations.Memberships.ListAutoPaging(
+			ctx,
+			cmd.Value("organization-id").(string),
+			params,
+			options...,
+		)
+		maxItems := int64(-1)
+		if cmd.IsSet("max-items") {
+			maxItems = cmd.Value("max-items").(int64)
+		}
+		return ShowJSONIterator(os.Stdout, "organizations:memberships list", iter, format, transform, maxItems)
+	}
+}
+
+func handleOrganizationsMembershipsGet(ctx context.Context, cmd *cli.Command) error {
+	client := nirvana.NewClient(getDefaultRequestOptions(cmd)...)
+	unusedArgs := cmd.Args().Slice()
+	if !cmd.IsSet("organization-id") && len(unusedArgs) > 0 {
+		cmd.Set("organization-id", unusedArgs[0])
+		unusedArgs = unusedArgs[1:]
+	}
+	if !cmd.IsSet("membership-id") && len(unusedArgs) > 0 {
+		cmd.Set("membership-id", unusedArgs[0])
+		unusedArgs = unusedArgs[1:]
+	}
+	if len(unusedArgs) > 0 {
+		return fmt.Errorf("Unexpected extra arguments: %v", unusedArgs)
+	}
+
+	options, err := flagOptions(
+		cmd,
+		apiquery.NestedQueryFormatBrackets,
+		apiquery.ArrayQueryFormatComma,
+		EmptyBody,
+		false,
+	)
+	if err != nil {
+		return err
+	}
+
+	var res []byte
+	options = append(options, option.WithResponseBodyInto(&res))
+	_, err = client.Organizations.Memberships.Get(
+		ctx,
+		cmd.Value("organization-id").(string),
+		cmd.Value("membership-id").(string),
+		options...,
+	)
+	if err != nil {
+		return err
+	}
+
+	obj := gjson.ParseBytes(res)
+	format := cmd.Root().String("format")
+	transform := cmd.Root().String("transform")
+	return ShowJSON(os.Stdout, "organizations:memberships get", obj, format, transform)
+}